requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.4.0
azure-core==1.22.1
azure-storage-blob==12.9.0
cachelib==0.6.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
email-validator==1.1.3
fake-useragent==0.1.11
Flask==2.0.3
Flask-Session==0.4.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.5
httplib2==0.20.2
httptools==0.2.0
httpx==0.21.3
idna==2.10
importlib-metadata==4.10.1
iniconfig==1.1.1
isodate==0.6.1
itsdangerous==2.0.1
Jinja2==3.0.3
requests==2.25.1
SQLAlchemy==1.4.31
sqlparse==0.4.2
urllib3==1.26.7
uvicorn==0.15.0
websockets==10.1
Werkzeug==2.0.3

